FLYW-131 refactor: 가격 산정 도메인 파일 구조 수정 (#78)
